test db & rename test repository
