+++ src/main/java/bank/recommendationservice/fintech/testrepository/RecommendationsRepository.java
@@ -1,4 +1,4 @@
-package bank.recommendationservice.fintech.repository;
+package bank.recommendationservice.fintech.testrepository;
 
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.jdbc.core.JdbcTemplate;
